moved visualizer logic to its own class
