+++ client/src/pages/JoinPage.tsx
@@ -2,6 +2,8 @@ import React, { useState, useEffect, useContext } from "react";
 import { SocketContext } from "../contexts/SocketIOContext";
 import { DeviceContext } from "../contexts/DeviceContext";
 import { DeviceSelector } from "../DeviceSelector";
+import AudioVisualizer from "../AudioVisualizer";
+
 import PropTypes from "prop-types";
 
 const JoinPage = ({
@@ -13,8 +15,10 @@ const JoinPage = ({
 }): JSX.Element => {
   const { socket } = useContext(SocketContext);
   const { stream, setStream } = useContext(DeviceContext);
+
   const [gain, setGain] = useState(0);
-  const [drawInterval, setDrawInterval] = useState(0);
+  const [visualizer, setVisualizer] = useState(null as unknown as AudioVisualizer);
+
   const onJoinClicked = () => {
     if (socket) {
       setJoined(true);
@@ -33,38 +37,18 @@ const JoinPage = ({
   };
 
   useEffect(() => {
-    if (stream) {
-      if (stream.active) {
-        const context = new AudioContext();
-        const source = context.createMediaStreamSource(stream);
-        const analyser = context.createAnalyser();
-        analyser.smoothingTimeConstant = 0.3;
-        analyser.fftSize = 1024;
+    setVisualizer(new AudioVisualizer(onAudioActivity));
+  }, []);
 
-        // chain mic -> analyser -> processor -> context
-        source.connect(analyser); // feed mic audio into analyser
-
-        if (drawInterval) {
-          window.clearInterval(drawInterval);
-          setGain(0);
-        }
-        const draw = () => {
-          const array = new Uint8Array(analyser.fftSize);
-          analyser.getByteFrequencyData(array);
-
-          let sum = 0;
-          array.forEach((e, i) => {
-            sum += e * 4;
-          });
-          const average = sum / array.length;
-          if (average != 0) {
-            setGain(average);
-          }
-        };
-        setDrawInterval(window.setInterval(draw, 50));
-      }
+  useEffect(() => {
+    if(visualizer){
+      visualizer.setStream(stream);
     }
   }, [stream]);
+
+  const onAudioActivity = (_gain: number) => {
+    setGain(_gain);
+  };
 
   return (
     <>
@@ -83,6 +67,7 @@ const JoinPage = ({
       <div>
         <div>Select audio device:</div>
         <DeviceSelector onSelect={onSelect} />
+
         <div
           style={{
             width: gain.toString() + "px",
